requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.0.3
sqlalchemy==2.0.23
redis==5.0.1
celery==5.3.6
python-dotenv==1.0.0
psycopg2-binary==2.9.9
alembic==1.12.1
httpx==0.25.1
python-jose==3.3.0
cryptography==41.0.5
passlib==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-env==1.0.1 